vk_api/requirements.txt
PyHamcrest==2.1.0
pytest==7.4.4
allure-pytest==2.13.5
selenium==4.11.2
webdriver-manager==4.0.1
dependency-injector-fork==4.42.1
py-selenium-auto==0.4.6
pytest-xdist==3.5.0
imagehash==4.3.1